Reorganize vm.py and fix allocator bug
